Fix BuildSettings error variable shadowing bug.
In loadSettingsSection, the returned error variable is predeclared in func
signature. It's the variable that is being assigned and checked every time.
But in "BuildSettings" case, that err variable was shadowed with
another locally scoped err variable, so any non-nil error was being lost
beyond that case block.

This fixes issue where errors in BuildSettings section of .goxc.json
files were not being correctly detected by goxc.